requirements.txt
# Core dependencies
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.31.0
torch>=2.5.0
torchvision>=0.20.0
opencv-python>=4.10.0.84

# Geometric processing
geopandas>=1.0.0
shapely>=2.0.6
fiona>=1.10.0
pyproj>=3.7.0

# Data processing
numpy>=2.1.0
pandas>=2.2.3
scikit-learn>=1.6.0

# Visualization
matplotlib>=3.9.2
seaborn>=0.13.2
pillow>=11.0.0
grad-cam>=1.5.0

# Reporting
reportlab>=4.2.5
python-multipart>=0.0.12

# Utilities
tqdm>=4.66.5
scipy>=1.14.1